requirements.txt
ansi2html==1.8.0
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
et-xmlfile==1.1.0
Flask==2.2.5
folium==0.14.0
fonttools==4.40.0
future==0.18.3
geocoder==1.38.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
psycopg2==2.9.7
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
ratelim==0.1.6
requests==2.31.0
retrying==1.3.4
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.20
tenacity==8.2.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.2.3